+++ packages/core/kind/tests/kind-specs.js
@@ -29,14 +29,14 @@ describe('kind', () => {
 		}
 	});
 
-	it('Should assign name to displayName', function () {
+	it('should assign name to displayName', function () {
 		const expected = 'Kind';
 		const actual = Kind.displayName;
 
 		expect(actual).to.equal(expected);
 	});
 
-	it('Should default {label} property', function () {
+	it('should default {label} property', function () {
 		const subject = <Kind prop={1} />;
 
 		const expected = 'Label';
@@ -45,7 +45,7 @@ describe('kind', () => {
 		expect(actual).to.equal(expected);
 	});
 
-	it('Should default {label} property when explicitly undefined', function () {
+	it('should default {label} property when explicitly undefined', function () {
 		// Explicitly testing for undefined
 		// eslint-disable-next-line no-undefined
 		const subject = <Kind prop={1} label={undefined} />;
@@ -56,7 +56,7 @@ describe('kind', () => {
 		expect(actual).to.equal(expected);
 	});
 
-	it('Should add className defined in styles', function () {
+	it('should add className defined in styles', function () {
 		const subject = mount(
 			<Kind prop={1} />
 		);
@@ -67,7 +67,7 @@ describe('kind', () => {
 		expect(actual).to.equal(expected);
 	});
 
-	it('Should compute {value} property', function () {
+	it('should compute {value} property', function () {
 		const subject = mount(
 			<Kind prop={1} />
 		);