+++ packages/sampler/stories/moonstone-stories/Picker.js
@@ -1,8 +1,9 @@
 import Picker, {PickerBase} from '@enact/moonstone/Picker';
 import Changeable from '@enact/ui/Changeable';
+import {decrementIcons, incrementIcons} from './icons';
 import React from 'react';
 import {storiesOf, action} from '@kadira/storybook';
-import {withKnobs, text, boolean, select} from '@kadira/storybook-addon-knobs';
+import {withKnobs, boolean, select} from '@kadira/storybook-addon-knobs';
 
 const StatefulPicker = Changeable(Picker);
 StatefulPicker.propTypes = Object.assign({}, PickerBase.propTypes, StatefulPicker.propTypes);
@@ -31,14 +32,14 @@ storiesOf('Picker')
 		() => (
 			<StatefulPicker
 				onChange={action('onChange')}
-				width={nullify(select('width', prop.width, 'large'))}
-				orientation={select('orientation', prop.orientation)}
-				wrap={boolean('wrap')}
-				joined={boolean('joined')}
-				noAnimation={boolean('noAnimation')}
-				disabled={boolean('disabled')}
-				incrementIcon={text('incrementIcon')}
-				decrementIcon={text('decrementIcon')}
+				width={nullify(select('width', prop.width, prop.width[3]))}
+				orientation={select('orientation', prop.orientation, prop.orientation[0])}
+				wrap={boolean('wrap', false)}
+				joined={boolean('joined', false)}
+				noAnimation={boolean('noAnimation', false)}
+				disabled={boolean('disabled', false)}
+				incrementIcon={select('incrementIcon', ['', ...incrementIcons])}
+				decrementIcon={select('decrementIcon', ['', ...decrementIcons])}
 			>
 				{airports}
 			</StatefulPicker>