+++ packages/sampler/stories/moonstone-stories/Popup.js
@@ -1,4 +1,5 @@
 import Popup, {PopupBase} from '@enact/moonstone/Popup';
+import BodyText from '@enact/moonstone/BodyText';
 import React from 'react';
 import {storiesOf, action} from '@kadira/storybook';
 import {withKnobs, boolean, text, select} from '@kadira/storybook-addon-knobs';
@@ -12,14 +13,17 @@ storiesOf('Popup')
 		' ',
 		'Basic usage of Popup',
 		() => (
-			<Popup
-				open={boolean('open', true)}
-				noAnimation={boolean('noAnimation', false)}
-				noAutoDismiss={boolean('noAutoDismiss', false)}
-				onClose={action('onClose')}
-				showCloseButton={boolean('showCloseButton', false)}
-				spotlightRestrict={select('spotlightRestrict', ['none', 'self-first', 'self-only'], 'self-only')}
-			>
-				<div>{text('children', 'Hello Popup')}</div>
-			</Popup>
+			<div>
+				<Popup
+					open={boolean('open', false)}
+					noAnimation={boolean('noAnimation', false)}
+					noAutoDismiss={boolean('noAutoDismiss', false)}
+					onClose={action('onClose')}
+					showCloseButton={boolean('showCloseButton', false)}
+					spotlightRestrict={select('spotlightRestrict', ['none', 'self-first', 'self-only'], 'self-only')}
+				>
+					<div>{text('children', 'Hello Popup')}</div>
+				</Popup>
+				<BodyText centered>Use KNOBS to interact with Popup.</BodyText>
+			</div>
 		));