+++ packages/sampler/stories/moonstone-stories/ProgressBar.js
@@ -14,10 +14,8 @@ storiesOf('ProgressBar')
 		'The basic ProgressBar',
 		() => (
 			<ProgressBar
-				backgroundProgress={number('backgroundProgress')}
-				max={number('max')}
-				min={number('min')}
-				progress={number('progress')}
+				backgroundProgress={number('backgroundProgress', ProgressBar.defaultProps.backgroundProgress, {range: true, min: 0, max: 1, step: 0.01})}
+				progress={number('progress', ProgressBar.defaultProps.progress, {range: true, min: 0, max: 1, step: 0.1})}
 				vertical={boolean('vertical')}
 				disabled={boolean('disabled')}
 			/>