+++ packages/spotlight/src/container.js
@@ -10,12 +10,19 @@ const spotlightDefaultClass = 'spottable-default';
 const enterEvent = 'onMouseEnter';
 const leaveEvent = 'onMouseLeave';
 
+/**
+ * Default config for {@link spotlight.SpotlightContainerDecorator}
+ *
+ * @memberof spotlight.SpotlightContainerDecorator
+ * @hocconfig
+ */
 const defaultConfig = {
 	/**
 	 * The selector for the default spottable element within the container.
 	 *
 	 * @type {String}
 	 * @default '.spottable-default'
+	 * @memberof spotlight.SpotlightContainerDecorator.defaultConfig
 	 * @public
 	 */
 	defaultElement: `.${spotlightDefaultClass}`,
@@ -25,6 +32,7 @@ const defaultConfig = {
 	 *
 	 * @type {String}
 	 * @default 'last-focused'
+	 * @memberof spotlight.SpotlightContainerDecorator.defaultConfig
 	 * @public
 	 */
 	enterTo: 'last-focused',
@@ -34,6 +42,7 @@ const defaultConfig = {
 	 *
 	 * @type {Boolean}
 	 * @default false
+	 * @memberof spotlight.SpotlightContainerDecorator.defaultConfig
 	 * @public
 	 */
 	preserveId: false
@@ -65,6 +74,9 @@ const defaultConfig = {
  * @param  {Function} Higher-order component
  *
  * @returns {Function} SpotlightContainerDecorator
+ * @class SpotlightContainerDecorator
+ * @memberof spotlight
+ * @hoc
  */
 const SpotlightContainerDecorator = hoc(defaultConfig, (config, Wrapped) => {
 	const forwardMouseEnter = forward(enterEvent);
@@ -74,7 +86,7 @@ const SpotlightContainerDecorator = hoc(defaultConfig, (config, Wrapped) => {
 	return class extends React.Component {
 		static displayName = 'SpotlightContainerDecorator';
 
-		static propTypes = /** @lends spotlight/SpotlightContainerDecorator.SpotlightContainerDecorator */ {
+		static propTypes = /** @lends spotlight.SpotlightContainerDecorator.protoype */ {
 			/**
 			 * Specifies the container id. If the value is `null`, an id will be generated.
 			 *
@@ -82,6 +94,15 @@ const SpotlightContainerDecorator = hoc(defaultConfig, (config, Wrapped) => {
 			 * @public
 			 */
 			containerId: PropTypes.string,
+
+			/**
+			 * When `true`, controls in the container cannot be navigated.
+			 *
+			 * @type {Boolean}
+			 * @default false
+			 * @public
+			 */
+			spotlightDisabled: PropTypes.bool,
 
 			/**
 			 * Whether or not the container is in muted mode.
@@ -94,7 +115,12 @@ const SpotlightContainerDecorator = hoc(defaultConfig, (config, Wrapped) => {
 
 			/**
 			 * Restricts or prioritizes navigation when focus attempts to leave the container. It
-			 * can be either 'none', 'self-first', or 'self-only'.
+			 * can be either 'none', 'self-first', or 'self-only'. Specifying 'self-first' indicates that
+			 * elements within the container will have a higher likelihood to be chosen as the next
+			 * navigable element. Specifying 'self-only' indicates that elements in other containers
+			 * cannot be navigated to by using 5-way navigation - however, elements in other containers
+			 * can still receive focus by calling `Spotlight.focus(elem)` explicitly. Specying 'none'
+			 * indicates there should be no restrictions when 5-way navigating the container.
 			 *
 			 * @type {String}
 			 * @default 'none'
@@ -104,6 +130,7 @@ const SpotlightContainerDecorator = hoc(defaultConfig, (config, Wrapped) => {
 		}
 
 		static defaultProps = {
+			spotlightDisabled: false,
 			spotlightMuted: false,
 			spotlightRestrict: 'none'
 		}
@@ -169,13 +196,17 @@ const SpotlightContainerDecorator = hoc(defaultConfig, (config, Wrapped) => {
 		}
 
 		render () {
-			const {spotlightMuted, ...rest} = this.props;
+			const {spotlightDisabled, spotlightMuted, ...rest} = this.props;
 			delete rest.containerId;
 			delete rest.spotlightRestrict;
 
 			rest['data-container-id'] = this.state.id;
 			rest[enterEvent] = this.handleMouseEnter;
 			rest[leaveEvent] = this.handleMouseLeave;
+
+			if (spotlightDisabled) {
+				rest['data-container-disabled'] = spotlightDisabled;
+			}
 
 			if (spotlightMuted) {
 				rest['data-container-muted'] = spotlightMuted;