+++ packages/ui/Holdable/Holdable.js
@@ -7,7 +7,7 @@
 import {off, once} from '@enact/core/dispatcher';
 import {forward} from '@enact/core/handle';
 import hoc from '@enact/core/hoc';
-import contains from 'ramda/src/contains';
+import {is} from '@enact/core/keymap';
 import pick from 'ramda/src/pick';
 import React, {PropTypes} from 'react';
 
@@ -34,13 +34,14 @@ const pointerEnter = 'onMouseEnter';
 const pointerLeave = 'onMouseLeave';
 const pointerMove = 'onMouseMove';
 
-const selectionKeyCodes = [13, 16777221];
+const isEnter = is('enter');
 
 /**
  * Default config for {@link ui/Holdable.Holdable}
  *
- * @memberof ui/Holdable
+ * @memberof ui/Holdable.Holdable
  * @hocconfig
+ * @public
  */
 const defaultConfig = {
 	/**
@@ -56,7 +57,7 @@ const defaultConfig = {
 	 *
 	 * @type {String}
 	 * @default 'onMove'
-	 * @memberof ui/Holdable.defaultConfig
+	 * @memberof ui/Holdable.Holdable.defaultConfig
 	 */
 	endHold: 'onMove',
 
@@ -78,7 +79,7 @@ const defaultConfig = {
 	 *
 	 * @type {Array}
 	 * @default [{name: 'hold', time: 200}]
-	 * @memberof ui/Holdable.defaultConfig
+	 * @memberof ui/Holdable.Holdable.defaultConfig
 	 */
 	events: [
 		{name: 'hold', time: 200}
@@ -92,7 +93,7 @@ const defaultConfig = {
 	 *
 	 * @type {Number}
 	 * @default 200
-	 * @memberof ui/Holdable.defaultConfig
+	 * @memberof ui/Holdable.Holdable.defaultConfig
 	 */
 	frequency: 200,
 
@@ -104,7 +105,7 @@ const defaultConfig = {
 	 *
 	 * @type {Number}
 	 * @default 16
-	 * @memberof ui/Holdable.defaultConfig
+	 * @memberof ui/Holdable.Holdable.defaultConfig
 	 */
 	moveTolerance: 16,
 
@@ -116,7 +117,7 @@ const defaultConfig = {
 	 *
 	 * @type {Boolean}
 	 * @default false
-	 * @memberof ui/Holdable.defaultConfig
+	 * @memberof ui/Holdable.Holdable.defaultConfig
 	 */
 	resume: false
 };
@@ -127,7 +128,7 @@ const defaultConfig = {
  *
  * @class Holdable
  * @memberof ui/Holdable
- * @ui
+ * @hoc
  * @public
  */
 const HoldableHOC = hoc(defaultConfig, (config, Wrapped) => {
@@ -140,7 +141,7 @@ const HoldableHOC = hoc(defaultConfig, (config, Wrapped) => {
 	const forwardPointerMove = forward(pointerMove);
 
 	return class Holdable extends React.Component {
-		static propTypes = /** @lends ui/Holdable.Holdable */ {
+		static propTypes = /** @lends ui/Holdable.Holdable.prototype */ {
 			/**
 			 * Whether or not the component is in a disabled state.
 			 *
@@ -193,7 +194,7 @@ const HoldableHOC = hoc(defaultConfig, (config, Wrapped) => {
 
 		onKeyDepress = (ev) => {
 			if (!this.props.disabled) {
-				if (contains(ev.keyCode, selectionKeyCodes) && !this.holdJob) {
+				if (isEnter(ev.keyCode) && !this.holdJob) {
 					this.keyEvent = true;
 					this.beginHold(pick(eventProps, ev));
 				}
@@ -202,7 +203,7 @@ const HoldableHOC = hoc(defaultConfig, (config, Wrapped) => {
 		}
 
 		onKeyRelease = (ev) => {
-			if (contains(ev.keyCode, selectionKeyCodes)) {
+			if (isEnter(ev.keyCode)) {
 				this.keyEvent = false;
 				this.endHold();
 			}