+++ packages/ui/Toggleable/Toggleable.js
@@ -12,7 +12,7 @@ import React from 'react';
 /**
  * Default config for {@link ui/Toggleable.Toggleable}
  *
- * @memberof ui/Toggleable
+ * @memberof ui/Toggleable.Toggleable
  * @hocconfig
  */
 const defaultConfig = {
@@ -21,7 +21,7 @@ const defaultConfig = {
 	 *
 	 * @type {Boolean}
 	 * @default false
-	 * @memberof ui/Toggleable.defaultConfig
+	 * @memberof ui/Toggleable.Toggleable.defaultConfig
 	 */
 	mutable: false,
 
@@ -29,7 +29,7 @@ const defaultConfig = {
 	 * Configures the event name that activates the component
 	 *
 	 * @type {String}
-	 * @memberof ui/Toggleable.defaultConfig
+	 * @memberof ui/Toggleable.Toggleable.defaultConfig
 	 */
 	activate: null,
 
@@ -37,7 +37,7 @@ const defaultConfig = {
 	 * Configures the event name that deactivates the component
 	 *
 	 * @type {String}
-	 * @memberof ui/Toggleable.defaultConfig
+	 * @memberof ui/Toggleable.Toggleable.defaultConfig
 	 */
 	deactivate: null,
 
@@ -46,7 +46,7 @@ const defaultConfig = {
 	 *
 	 * @type {String}
 	 * @default 'onToggle'
-	 * @memberof ui/Toggleable.defaultConfig
+	 * @memberof ui/Toggleable.Toggleable.defaultConfig
 	 */
 	toggle: 'onToggle',
 
@@ -55,7 +55,7 @@ const defaultConfig = {
 	 *
 	 * @type {String}
 	 * @default 'active'
-	 * @memberof ui/Toggleable.defaultConfig
+	 * @memberof ui/Toggleable.Toggleable.defaultConfig
 	 */
 	prop: 'active'
 };
@@ -68,7 +68,7 @@ const defaultConfig = {
  *
  * @class Toggleable
  * @memberof ui/Toggleable
- * @ui
+ * @hoc
  * @public
  */
 const ToggleableHOC = hoc(defaultConfig, (config, Wrapped) => {
@@ -112,7 +112,7 @@ const ToggleableHOC = hoc(defaultConfig, (config, Wrapped) => {
 		}
 
 		componentWillReceiveProps (nextProps) {
-			const active = nextProps[prop];
+			const active = !!nextProps[prop];
 			if (mutable && this.props[prop] !== active) {
 				this.setState({active});
 			}